컴포넌트 루트명 및 파일명 변경, 검색 컴포넌트 생성(작업중)
